Finished and refactored MobileNavbar to Router
The mobile navbar is finished. You can view it at "/MobileNav"  when hosting the server.
To view on your mobile device, go to vite.config.ts and after the plugins key in the define function, put

server: {
    host: true,
    port: 3000,
  }

and then visit the page on your mobile device using the ip shown in the console.

To integrate the navbar onto other pages, make the route element of the page a child of the navbar's route element (In between the tags).

The only thing left to do is to Link the buttons to the correct pages once they're created.