READ THIS COMMIT
used a new line as an excuse to commit.

I've realized that the routes is not the location where this sort of implementation should be, it's in controllers. However, I have no reference to go by and I'm not experienced with backend development like this. Once I can see one reference of controller, and a proper reference of what routes can/should look like, I can update profiles. Until then, it will be in routes.